Adding separate jet clustering helper class
